Address compiler warnings
1) Signed/unsigned variables. Rationale 1: when comparing int to unsigned int
when possible/sensible, change the unsigned int to a larger, signed int
Rationale 2: for loops over size() use a size_t counter
2) Remove unused variables
3) Limit quick game txs to valid quick game type numbers (0x00)
4) Give default value of 0 to CMappingTX int vars to make switch checks on undefined values impossible

+++ src/betting/bet_common.cpp
@@ -172,7 +172,7 @@ void GetPLRewardPayouts(const uint32_t nNewBlockHeight, std::vector<CBetOut>& vE
     CScript payoutScriptOMNO = GetScriptForDestination(CBitcoinAddress(Params().OMNOPayoutAddr()).Get());
 
     // Loop over the payout vector and aggregate values.
-    for (int i = 0; i < vExpectedPayouts.size(); i++)
+    for (size_t i = 0; i < vExpectedPayouts.size(); i++)
     {
         if (vPayoutsInfo[i].payoutType == PayoutType::bettingPayout) {
             CAmount betValue = vExpectedPayouts[i].nBetValue;
